requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==0.8.11
pillow==10.1.0
requests==2.31.0
pydantic==2.4.2
